backend/requirements.txt
asgiref==3.11.0
black==25.11.0
click==8.1.8
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.4.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytokens==0.4.0
sqlparse==0.5.5
tomli==2.4.0
typing_extensions==4.15.0
whitenoise==6.11.0
